Updated configuration and reduced code
