+++ frameworks/Java/activeweb/src/main/webapp/WEB-INF/views/fortunes/index.html
@@ -1,20 +1,1 @@
-<!DOCTYPE html>
-<html>
-<head><title>Fortunes</title></head>
-<body>
-<table>
-    <tr>
-        <th>id</th>
-        <th>message</th>
-    </tr>
-
-<#list fortunes as fortune >
-    <tr>
-        <td>${fortune.id}</td>
-        <td>${fortune.message}</td>
-    </tr>
-</#list>
-
-</table>
-</body>
-</html>
+<!DOCTYPE html><html><head><title>Fortunes</title></head><body><table><tr><th>id</th><th>message</th></tr><#list fortunes as fortune ><tr><td>${fortune.id}</td><td>${fortune.message}</td></tr></#list></table></body></html>
