+++ frameworks/Java/activeweb/src/main/webapp/WEB-INF/views/queries/index.html
@@ -1,5 +1,1 @@
-[<#list worlds as w >
-    {"id":${w.id},"randomNumber":${w.randomNumber}}
-    <#if w_has_next >,</#if>
-</#list>]
-
+[<#list worlds as w >{"id":${w.id},"randomNumber":${w.randomNumber}}<#if w_has_next >,</#if></#list>]
